+++ src/commands/changeDelay-4.lua
@@ -3,12 +3,11 @@
   Input:
     KEYS[1] delayed key
     KEYS[2] meta key
-    KEYS[3] active key
-    KEYS[4] marker key
-    KEYS[5] events stream
+    KEYS[3] marker key
+    KEYS[4] events stream
 
     ARGV[1] delay
-    ARGV[2] delayedTimestamp
+    ARGV[2] timestamp
     ARGV[3] the id of the job
     ARGV[4] job key
 
@@ -24,13 +23,14 @@ local rcall = redis.call
 
 -- Includes
 --- @include "includes/addDelayMarkerIfNeeded"
+--- @include "includes/getDelayedScore"
 --- @include "includes/getOrSetMaxEvents"
---- @include "includes/isQueuePausedOrMaxed"
 
 if rcall("EXISTS", ARGV[4]) == 1 then
   local jobId = ARGV[3]
-  local score = tonumber(ARGV[2])
-  local delayedTimestamp = (score / 0x1000)
+
+  local delay = tonumber(ARGV[1])
+  local score, delayedTimestamp = getDelayedScore(KEYS[1], ARGV[2], delay)
 
   local numRemovedElements = rcall("ZREM", KEYS[1], jobId)
 
@@ -38,19 +38,16 @@ if rcall("EXISTS", ARGV[4]) == 1 then
     return -3
   end
 
-  rcall("HSET", ARGV[4], "delay", tonumber(ARGV[1]))
+  rcall("HSET", ARGV[4], "delay", delay)
   rcall("ZADD", KEYS[1], score, jobId)
 
   local maxEvents = getOrSetMaxEvents(KEYS[2])
 
-  rcall("XADD", KEYS[5], "MAXLEN", "~", maxEvents, "*", "event", "delayed",
+  rcall("XADD", KEYS[4], "MAXLEN", "~", maxEvents, "*", "event", "delayed",
     "jobId", jobId, "delay", delayedTimestamp)
 
   -- mark that a delayed job is available
-  local isPausedOrMaxed = isQueuePausedOrMaxed(KEYS[2], KEYS[3])
-  if not isPausedOrMaxed then
-    addDelayMarkerIfNeeded(KEYS[4], KEYS[1])
-  end
+  addDelayMarkerIfNeeded(KEYS[3], KEYS[1])
 
   return 0
 else