+++ src/commands/includes/getNextDelayedTimestamp.lua
@@ -5,9 +5,8 @@ local function getNextDelayedTimestamp(delayedKey)
   local result = rcall("ZRANGE", delayedKey, 0, 0, "WITHSCORES")
   if #result then
     local nextTimestamp = tonumber(result[2])
-    if nextTimestamp ~= nil then
-      nextTimestamp = nextTimestamp / 0x1000
+    if nextTimestamp ~= nil then 
+      return nextTimestamp / 0x1000
     end
-    return nextTimestamp
   end
 end