+++ src/commands/moveToDelayed-8.lua
@@ -13,11 +13,10 @@
 
     ARGV[1] key prefix
     ARGV[2] timestamp
-    ARGV[3] delayedTimestamp
-    ARGV[4] the id of the job
-    ARGV[5] queue token
-    ARGV[6] delay value
-    ARGV[7] skip attempt
+    ARGV[3] the id of the job
+    ARGV[4] queue token
+    ARGV[5] delay value
+    ARGV[6] skip attempt
 
   Output:
     0 - OK
@@ -31,32 +30,32 @@ local rcall = redis.call
 
 -- Includes
 --- @include "includes/addDelayMarkerIfNeeded"
+--- @include "includes/getDelayedScore"
 --- @include "includes/getOrSetMaxEvents"
---- @include "includes/isQueuePausedOrMaxed"
 --- @include "includes/removeLock"
 
 local jobKey = KEYS[5]
 local metaKey = KEYS[7]
-local token = ARGV[5] 
+local token = ARGV[4] 
 if rcall("EXISTS", jobKey) == 1 then
-    local errorCode = removeLock(jobKey, KEYS[8], token, ARGV[4])
+    local errorCode = removeLock(jobKey, KEYS[8], token, ARGV[3])
     if errorCode < 0 then
         return errorCode
     end
 
     local delayedKey = KEYS[4]
-    local jobId = ARGV[4]
-    local score = tonumber(ARGV[3])
-    local delayedTimestamp = (score / 0x1000)
+    local jobId = ARGV[3]
+    local delay = tonumber(ARGV[5])
+    local score, delayedTimestamp = getDelayedScore(delayedKey, ARGV[2], delay)
 
     local numRemovedElements = rcall("LREM", KEYS[2], -1, jobId)
     if numRemovedElements < 1 then return -3 end
 
-    if ARGV[7] == "0" then
+    if ARGV[6] == "0" then
         rcall("HINCRBY", jobKey, "atm", 1)
     end
 
-    rcall("HSET", jobKey, "delay", ARGV[6])
+    rcall("HSET", jobKey, "delay", ARGV[5])
 
     local maxEvents = getOrSetMaxEvents(metaKey)
 
@@ -65,11 +64,8 @@ if rcall("EXISTS", jobKey) == 1 then
           "jobId", jobId, "delay", delayedTimestamp)
 
     -- Check if we need to push a marker job to wake up sleeping workers.
-    local isPausedOrMaxed = isQueuePausedOrMaxed(metaKey, KEYS[2])
-    if not isPausedOrMaxed then
-        local markerKey = KEYS[1]
-        addDelayMarkerIfNeeded(markerKey, delayedKey)
-    end
+    local markerKey = KEYS[1]
+    addDelayMarkerIfNeeded(markerKey, delayedKey)
 
     return 0
 else